LPS-112372 Move InfoForm to its own package
Thanks @4lejandrito for the suggestion
