+++ modules/apps/info/info-api/src/main/java/com/liferay/info/item/provider/InfoItemFormProvider.java
@@ -15,7 +15,7 @@
 package com.liferay.info.item.provider;
 
 import com.liferay.info.field.InfoFieldValue;
-import com.liferay.info.field.InfoForm;
+import com.liferay.info.form.InfoForm;
 import com.liferay.info.field.InfoFormValues;
 import com.liferay.info.item.NoSuchClassTypeException;
